refactor: 여행홈 app내부에서는 라우팅만 담당하도록 수정
