+++ snapsplit/src/features/trip/[tripId]/budget/type.ts
@@ -1,5 +1,9 @@
 import { Expense } from "./api";
 
+export type BudgetPageProps = {
+  tripId: string;
+};
+
 // TripInfo.tsx
 export type TripInfoProps = {
     tripName: string;
